requirements.txt
streamlit
pandas
python-dotenv
google-generativeai
faiss-cpu
PyPDF2
pydantic
